moral-consistency-llm/requirements.txt
# Gestione Ambiente e Variabili
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM & AI Clients
openai>=1.0.0
anthropic>=0.15.0
transformers>=4.38.0
torch>=2.2.0  # Se usi Mac M1/M2/M3, verrà installato con supporto MPS
huggingface_hub>=0.20.0

# Analisi Dati
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0   # Per la cosine similarity
sentence-transformers>=2.5.0 # Per calcolare gli embedding

# Testing e Utilities
pytest>=8.0.0
tqdm>=4.66.0  # Barra di caricamento per gli esperimenti